+++ graspit-reporters/packages/wdio-graspit-reporter/src/contacts.ts
@@ -15,6 +15,8 @@ log4js.configure({
 export default class ReporterContacts extends WDIOReporter {
   options: ReporterOptions;
 
+  skipTestRun: boolean = false;
+
   supporter: ReporterDialPad;
 
   logger: Logger;