+++ graspit/services/CommandLine/_init.py
@@ -1,0 +1,6 @@
+from click import group
+
+
+@group(name="Graspit")
+def handle_cli():
+    pass