+++ graspit/services/DBService/models/__init__.py
@@ -1,5 +1,6 @@
 from .result_base import SessionBase, SuiteBase, RunBase
-from .config_base import AttachmentBase, TestConfigBase, ConfigBase, ExportBase
+from .config_base import ConfigBase, ExportBase
+from .static_base import AttachmentBase, TestConfigBase, StaticBase
 from .dynamic_base import TaskBase, DynamicVideoBase, PrunedBase
 
 __all__ = [
@@ -13,4 +14,5 @@ __all__ = [
     PrunedBase,
     ExportBase,
     ConfigBase,
+    StaticBase,
 ]