+++ graspit/services/DBService/models/dynamic_base.py
@@ -1,8 +1,16 @@
-from graspit.services.DBService.models.config_base import AttachmentFields
+from graspit.services.DBService.models.static_base import AttachmentFields
 from graspit.services.DBService.models.result_base import SuiteBase
 from tortoise.models import Model
-from tortoise.fields import CharField, DatetimeField, JSONField, CharEnumField, ForeignKeyField, ForeignKeyRelation, \
-    BooleanField, TextField
+from tortoise.fields import (
+    CharField,
+    DatetimeField,
+    JSONField,
+    CharEnumField,
+    ForeignKeyField,
+    ForeignKeyRelation,
+    BooleanField,
+    TextField,
+)
 from graspit.services.SchedularService.constants import JobType
 from graspit.services.DBService.models.result_base import RunBase
 from graspit.services.DBService.models.enums import PrunedRecords
@@ -15,11 +23,19 @@ class TaskBase(Model):
     type = CharEnumField(JobType, null=False)
     dropped = DatetimeField(auto_now=True)  # use modified timestamp
     # this would schedule the parent suites in the later rounds
-    meta = JSONField(null=True, default={}, description="Data required to process the task, Not used as of now though")
+    meta = JSONField(
+        null=True,
+        default={},
+        description="Data required to process the task, Not used as of now though",
+    )
     test: ForeignKeyRelation[RunBase] = ForeignKeyField(
         "models.RunBase", related_name="tasks", to_field="testID"
     )
-    picked = BooleanField(null=True, default=False, description="True if the task is picked by the job else False")
+    picked = BooleanField(
+        null=True,
+        default=False,
+        description="True if the task is picked by the job else False",
+    )
 
 
 class DynamicVideoBase(AttachmentFields):
@@ -30,7 +46,11 @@ class DynamicVideoBase(AttachmentFields):
 
 
 class PrunedBase:
-    prunedID = CharField(max_length=36, description="possibly uuid of length 36 that was pruned")
-    reason = TextField(description='Reason for its existence here')
-    details = JSONField(null=True, default={}, description="If any details were required")
+    prunedID = CharField(
+        max_length=36, description="possibly uuid of length 36 that was pruned"
+    )
+    reason = TextField(description="Reason for its existence here")
+    details = JSONField(
+        null=True, default={}, description="If any details were required"
+    )
     type = CharEnumField(PrunedRecords, null=False)