+++ graspit/services/SchedularService/completeTestRun.py
@@ -3,12 +3,15 @@ from graspit.services.DBService.models.result_base import (
     RunBase,
     SuiteBase,
 )
+from uuid import UUID
+from typing import Union
+from traceback import format_exc
+from graspit.services.SchedularService.register import mark_for_prune_task
 from graspit.services.DBService.models.dynamic_base import TaskBase
 from graspit.services.DBService.models.types import Status, SuiteType
 from graspit.services.SchedularService.refer_types import PathTree, PathItem
 from graspit.services.SchedularService.modifySuites import fetch_key_from_status
-from graspit.services.SchedularService.pruneTasks import pruneTasks
-from graspit.services.DBService.models.config_base import TestConfigBase, AttachmentType
+from graspit.services.DBService.models.static_base import TestConfigBase, AttachmentType
 from tortoise.functions import Sum, Max, Min, Count, Lower
 from datetime import datetime
 from typing import List
@@ -85,38 +88,26 @@ def simplify_file_paths(paths: List[str]):
     return tree
 
 
-async def patchTestRun(test_id: str, current_test_id: str):
-    logger.info("Patching up the test run... {}", test_id)
+async def skip_test_run(test_id: Union[str, UUID], reason: str, **extra) -> False:
+    logger.error(reason)
+    await TestConfigBase.create(
+        test_id=str(test_id),
+        attachmentValue=dict(reason=reason, test_id=str(test_id), **extra),
+        type=AttachmentType.ERROR,
+        description="Job Failed: Complete Test Run",
+    )
+    await mark_for_prune_task(test_id)
+    return False
 
-    test_run = await RunBase.filter(testID=test_id).first()
-    task = await TaskBase.filter(ticketID=test_run.testID).first()
 
-    if test_run.standing != Status.PENDING:
-        logger.warning("Invalid task for completing the test run | {}", test_id)
-        return await task.delete()
+async def patchValues(test_run: RunBase):
+    test_id = test_run.testID
 
     filtered = SessionBase.filter(test_id=test_id)
-
-    pending_items = (
-        await SuiteBase.filter(session__test_id=test_id)
-        .filter(Q(standing=Status.YET_TO_CALCULATE) | Q(standing=Status.PENDING))
-        .count()
-    )
-
-    if pending_items > 0:
-        if await mark_test_failure_if_required(test_id):
-            return
-        logger.warning(
-            "Fix test run: {} was picked but some of its test entities were not processed yet.",
-            test_id,
-        )
-        await task.update_from_dict(dict(picked=False))
-        return await task.save()  # continue in the next run
-
     if await filtered.count() == 0:
-        logger.warning("Detected test run: {} with no sessions", test_id)
-        await task.delete()
-        return await test_run.delete()
+        reason = f"Detected test run: {test_run.projectName} - {test_id} with no sessions, Hence skipping this."
+        await skip_test_run(test_id, reason)
+        return
 
     test_result = (
         await filtered.annotate(
@@ -139,6 +130,7 @@ async def patchTestRun(test_id: str, current_test_id: str):
             "actual_start",
         )
     )
+
     passed = test_result.get("total_passed", 0)
     failed = test_result.get("total_failed", 0)
     skipped = test_result.get("total_skipped", 0)
@@ -182,26 +174,59 @@ async def patchTestRun(test_id: str, current_test_id: str):
         )
     )
 
-    await test_run.save()
-    logger.info("Completed the patch for test run | {}", test_id)
-    return await task.delete()
+
+async def patchTestRun(test_id: str, current_test_id: str):
+    test_run = await RunBase.filter(testID=test_id).first()
+    if not test_run:
+        return
+    task = await TaskBase.filter(ticketID=test_run.testID).first()
+    if not task:
+        return
+
+    logger.info("Patching up the Test Run: {}", test_id)
+    if test_run.standing != Status.PENDING:
+        reason = "Expected Test Run: {} to be pending but found: {}"
+        return skip_test_run(test_id, reason)
+
+    pending_items = (
+        await SuiteBase.filter(session__test_id=test_id)
+        .filter(Q(standing=Status.YET_TO_CALCULATE) | Q(standing=Status.PENDING))
+        .count()
+    )
+
+    if pending_items > 0:
+        if await mark_test_failure_if_required(test_id):
+            return
+        logger.warning(
+            "Fix test run: {} was picked but some of its test entities were not processed yet.",
+            test_id,
+        )
+        await task.update_from_dict(dict(picked=False))
+        await task.save()  # continue in the next run
+        return False
+
+    to_return = False
+    try:
+        await patchValues(test_run)
+    except Exception:
+        reason = f"Failed to patch the test run, error in calculation, {format_exc()}"
+        await skip_test_run(test_id, reason)
+    else:
+        logger.info("Completed the patch for test run: {}", test_id)
+        to_return = True
+    finally:
+        await test_run.save()
+    return to_return
 
 
 async def mark_test_failure_if_required(test_id: str) -> bool:
     test_run = await RunBase.filter(testID=test_id).first()
     if test_run.ended is None:
-        await TestConfigBase.create(
-            test_id=test_id,
-            attachmentValue=dict(
-                reason=f"Failed to process {test_id}: {test_run.projectName} because it didn't end date",
-                incomplete=test_run.projectName,
-                test_id=test_id,
-            ),
-            type=AttachmentType.ERROR,
-            description="Job Failed: Complete Test Run",
+        await skip_test_run(
+            test_id,
+            f"Failed to process test Run{test_id}: {test_run.projectName} because it didn't have a end date",
+            incomplete=test_run.projectName,
         )
-
-        await pruneTasks()
         return True
 
     suites = [
@@ -217,15 +242,6 @@ async def mark_test_failure_if_required(test_id: str) -> bool:
         # it means there are child tasks ensured to complete the child suites
         return False
 
-    await TestConfigBase.create(
-        test_id=test_id,
-        attachmentValue=dict(
-            reason=f"Failed to process {test_id} because of incomplete child suites",
-            incomplete=suites,
-            test_id=test_id,
-        ),
-        type=AttachmentType.ERROR,
-        description="Job Failed: Complete Test Run",
-    )
-
-    await pruneTasks()
+    reason = f"Failed to process {test_id} because of incomplete child suites"
+    await skip_test_run(test_id, reason, incomplete=suites)
+    return True