+++ graspit/services/SchedularService/pruneTasks.py
@@ -1,17 +1,29 @@
 from graspit.services.DBService.models.enums import AttachmentType
-from graspit.services.DBService.models.config_base import TestConfigBase
-from graspit.services.DBService.models.dynamic_base import TaskBase
+from graspit.services.DBService.models.static_base import TestConfigBase
+from graspit.services.DBService.models.dynamic_base import TaskBase, JobType
+from loguru import logger
+from typing import Optional
+from tortoise.expressions import Q
 
 
-async def pruneTasks():
-    multiple_test_ids = await TaskBase.all().only("test_id").distinct().values_list("test_id", flat=True)
+async def pruneTasks(request_id: Optional[str] = ""):
+    if request_id:
+        logger.error("Deleting Few Tasks as per request")
+    else:
+        logger.warning("Pruning some Tasks")
 
-    for test_id in multiple_test_ids:
-        has_error = await TestConfigBase.exists(type=AttachmentType.ERROR, test_id=test_id)
-        if not has_error:
-            continue
-
-        await TaskBase.filter(test_id=test_id).delete()
-        # you can drop the tasks that would lead to deadlock
-        # since these tasks cannot be completed (as they have incomplete information)
-
+    await TaskBase.filter(
+        Q(
+            test_id__in=await TestConfigBase.filter(type=AttachmentType.ERROR)
+            .filter(
+                test_id__in=await TaskBase.all()
+                .only("test_id")
+                .distinct()
+                .values_list("test_id", flat=True)
+            )
+            .only("test_id")
+            .distinct()
+            .values_list("test_id", flat=True)
+        )
+        & ~Q(ticketID=request_id)
+    ).delete()